REFACTOR : 이넘 파일 위치 변경
